requirements.txt

torch==2.6.0
torchvision==0.21.0
torchtext==0.18.0

transformers==4.48.1
fair-esm==2.0.0
esm==3.2.1.post1
sentencepiece==0.2.0

numpy==1.26.4
pandas==2.3.2
scikit-learn==1.2.2

tqdm==4.67.1
einops==0.8.1
PyYAML==6.0.2
biopython==1.85

matplotlib==3.10.5
seaborn==0.13.2
